More work on account pref: show ratings
