+++ lib/src/view/home/home_tab_screen.dart
@@ -25,6 +25,7 @@ import 'package:lichess_mobile/src/model/puzzle/puzzle_providers.dart';
 import 'package:lichess_mobile/src/model/user/user_repository_providers.dart';
 import 'package:lichess_mobile/src/model/user/user.dart';
 import 'package:lichess_mobile/src/model/settings/play_preferences.dart';
+import 'package:lichess_mobile/src/view/account/show_rating_aware_widget.dart';
 import 'package:lichess_mobile/src/view/auth/sign_in_widget.dart';
 import 'package:lichess_mobile/src/view/puzzle/puzzle_screen.dart';
 import 'package:lichess_mobile/src/view/user/leaderboard_widget.dart';
@@ -304,7 +305,7 @@ class _HomeBody extends ConsumerWidget {
                     const _CreateAGame(),
                     const _DailyPuzzle(),
                     if (session != null) RecentGames(user: session.user),
-                    LeaderboardWidget(),
+                    ShowRatingPrefAwareWidget(child: LeaderboardWidget()),
                   ],
                 )
               : SliverList(
@@ -312,7 +313,7 @@ class _HomeBody extends ConsumerWidget {
                     const _CreateAGame(),
                     const _DailyPuzzle(),
                     if (session != null) RecentGames(user: session.user),
-                    LeaderboardWidget(),
+                    ShowRatingPrefAwareWidget(child: LeaderboardWidget()),
                   ]),
                 );
         } else {
