+++ test/model/account/account_repository_test.dart
@@ -60,6 +60,7 @@ void main() {
     "insightShare": 1,
     "keyboardMove": 0,
     "zen": 0,
+    "ratings": 1,
     "moveEvent": 2,
     "rookCastle": 1
   },
